pyproject.toml
[project]
name = "phil-docstrings"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "attrs>=25.4.0",
    "click>=8.3.1",
    "google-generativeai>=0.8.5",
]


